add Strip Clock effect
Similar to the Analog Clock overlay. However, it doesn't show the
current second, but adds a trail to the minute.

The trail and 5-minute marks use the current palette or the first color,
the current minute uses the second color, the current hour the third
color.

The effect works best, when the number of LEDs is a multiple of 60,
otherwise black borders are added the the from and end.