Update PIR Switch Usermod
I have added Home Assistant Discovery option (toggleable in Usermod screen).

I also created an Occupancy timer that is only used for external reporting.

Future enhancement could be it include the occupancy timer in the JSON reporting display and being able to turn the occupancy sensor on/off.

I was testing this for around 6 months and it was working well.  I've updated it today to replace changes made to the upstream version since.

I amended the MQTT publish procedure but have, so far as I can see, preserved the original functionality.