Add information about which branch is being
Add the information about which software branch, beta or stable, is being used on this device.
This is accessible through both the JSON API and MQTT, just like the version string.
This will automatically detect if it is a beta or not based on the release version configured in package.json to make sure it is always set properly.

This could be really useful for any consumer of the API that wants to make sure to check the correct channel to know potentially if a device is using the most recent version, for example and also to match it to the correct branch (beta, soundreactive, ...).