reduce memory needs of popcorn effect
On a matrix with 52 columns, popcorn was requesting around 30Kb of segment data.
This patch reduces the data to the actually necessary amount based on the "intensity" slider. If intensity is increased, it means that the effect will get a bigger chunk of data allocated - zero'd out but this does not hurt much.