platform.ini cleanup + version bumps
- make formatting more uniform
- organize environments hierarchically
- remove duplications
- upgrade platform versions
- remove some obosolete dependencies
- upgrade some dependencies while at it
- upgrade c++ version
- remove debug symbols from the esp32 release builds

Regarging platvorm version upgrades:
Most environments were using relatively recent platform versions, but the
base esp32 configuration was not. This does increase image size, but that
is a problem we have anyway, because we have envs that use more recent
platform versions. This change just results in fewer differences between
environments. But this change however increases the minimum gcc version the
code has to compile with, therefore enabling better features and more
recent c++ versions.

Regarding dependencies:
The original LittleFS dependency was removed, because all platforms are now
recent eneough to include it anyway.
The other removed dependency is
https://github.com/Aircoookie/arduino-esp32.git#1.0.6.4 because it is
incompatible with the more recent platform version. I could however not
figure out what it was good for originally, so hopefully this doesn't break
anything...