Adding dithering to PWM and proper phase shifting (draft)
Using dithering allows for longer pulse times if not FET driver is used. In current configuration, on the slowest speed setting the minimum on-time is 400ns, which is enough time to drive a FET directly from a GPIO.
The down side of using dithering: the PWM frequency drops for the lowest duty-cycle values (minimum frequency is basefrequency/(1<<ditheringbits)
phase shifting is added to distribute the load better over one period. the algorithm used is very simple and does not perfectly distribute the signals in all cases but it is not too bad for a nice-to-have feature.
TODO:
-dithering is only tested with 4bits, less may be possible
-the way the frequency settings are used is just a POC
-need a proper way to enable/disable phase shift as well as dithering