Perceptual brightness usermod
This usermod adds API commands for changing the global brightness in
steps that look even to the human eye, but in reality are small steps at
the low end and large steps at the high end. There are commands included
to put brightness controls on a single button, ramping the brightness
while the button is held down.